requirements.txt
pandas
numpy
torch
fastparquet
boto3
python-dotenv
pyarrow
wandb
scikit-learn
datasets
tqdm
datasketch
transformers[torch]
sentence-transformers
pyyaml
click
eng
matplotlib
seaborn
nltk
requests
wordcloud
unbabel-comet
evaluate